### 2019-07-11 #1561 @whedon
[PRE REVIEW]: X.509 Compliant Hybrid Certificates for the Post-Quantum Transition
**Submitting author:** @j-braun (<a href="http://orcid.org/0000-0001-6491-2250">Johannes Braun</a>)
**Repository:** <a href="https://github.com/CROSSINGTUD/bc-hybrid-certificates" target ="_blank">https://github.com/CROSSINGTUD/bc-hybrid-certificates</a>
**Version:** v1.0.0
