### 2019-07-24 #1561 @j-braun
@arfon @danielskatz - I am sorry for the delayed answer. I was busy the past days and did not have the possibility to track the issue all the time. So here is the answer to your question in regard to research applictions:
Our software was designed to be used by (post-quantum) cryptography researchers and IT security practitioners. It’s goal is to enable researchers to extensively test their newly designed schemes in real world environments and to demonstrate their practicality. This is a very important aspect on that field of research as it allows insights and gives valuable hints in regard to optimization and parameter setting. A prior mock-up implementation was for example used  for the following publication, which we also reference in the paper:
“N. Bindel, U. Herath, M. McKague, and D. Stebila. Transitioning to a quantum resistant public key infrastructure. In T. Lange and T. Takagi, editors, Post-Quantum Cryptography, pages 384{405, Cham, 2017. Springer International Publishing.”
The mock up implementation was used to do a first evaluation of the hybrid certificate approach in conjunction with the Q-TESLA scheme designed by Bindel at al. Our software now allows a more thorough evaluation of such schemes.

Should we explicitly add this to the paper?
