### 2019-07-24 0b07f2a
updated paper.bib, paper.pdf (added years to bibliography)
