+++ paper/paper.bib
@@ -33,17 +33,20 @@ isbn="978-3-319-59879-6"
   title = {{Open Quantum Safe - software for prototyping
 quantum-resistant cryptography}},
   howpublished = {\url{https://openquantumsafe.org/}},
+	year="2019",
   note = {Online; accessed 2019-07}
 }
 
 @misc{opensslHybridCert,
   title = {{Hybrid Certificates - C, OpenSSL integration}},
+	year="2019",
   howpublished = {\url{https://github.com/CROSSINGTUD/openssl-hybrid-certificates}},
    note = {Online; accessed 2019-07}
 }
 
 @misc{BCHybridCert,
   title = {{Hybrid Certificates - Java, Bouncy Castle integration}},
+	year="2019",
   howpublished = {\url{https://github.com/CROSSINGTUD/bc-hybrid-certificates}},
   note = {Online; accessed 2019-07}
 }
@@ -51,12 +54,14 @@ quantum-resistant cryptography}},
 @misc{bouncyCastle,
   title = {{The Legion of the Bouncy Castle}},
   howpublished = {\url{https://www.bouncycastle.org/}},
+	year="2013",
    note = {Online; accessed 2019-07}
 }
 
 @misc{tecDocu,
 	author ={Luca Gladiator},
   title = {{Hybrid Certificates for OpenSSL}},
+	year="2019",
   howpublished = {\url{https://github.com/CROSSINGTUD/openssl-hybrid-certificates/blob/OQS-OpenSSL_1_1_1-stable/HybridCert_technical_documentation.pdf}},
   note = {Online; accessed 2019-07}
 }
