### 2019-07-24 #1561 @danielskatz
@j-braun - we have rotating associate editors-in-chief, which is why two of us have asked about this, and this week, we'll let a third (@kyleniemeyer) reply and follow-up...
