### 2019-07-26 #1561 @kyleniemeyer
Hi @j-braun, thanks for your explanation. Yes, please do add some description of the research applications of the software to your paper; this may be most helpful near the beginning.
