### 2019-07-26 #1561 @kyleniemeyer
@danielskatz can you serve as editor for this submission?
