### 2019-07-26 #1561 @danielskatz
👋 @gcdeshpande & @trp07 - would you be willing to review this JOSS submission?
