### 2019-07-26 #1561 @danielskatz
👋 @jbasney - would you be willing to review this JOSS submission?  Or perhaps suggest others who would be well-suited.  See the [paper above](https://github.com/openjournals/joss-papers/blob/joss.01561/joss.01561/10.21105.joss.01561.pdf) and [info on JOSS reviewing](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)
