+++ paper/paper.md
@@ -44,6 +44,8 @@ However, to facilitate the transition, also the cryptographic infrastructure mus
 hybrids--multiple algorithms in parallel that are combined such that the hybrid scheme is secure as long as at least one of the parallely used algorithms is secure. 
 For the post-quantum transition a classical scheme is combined with a post-quantum scheme. This has two clear advantages compared to a direct switch to post-quantum secure algorithms: "hedging our bets" when the security of newer algorithms is not yet certain but the security of older primitives is already in question; and to achieve security and functionality both in post-quantum-aware and in a backwards-compatible way with not-yet-upgraded software.
 
+Our implementation has many applications in regard to evaluation and testing of new cryptographic schemes. It enables (post-quantum) cryptography researchers and IT security practitioners to extensively test newly designed schemes in real world environments and to demonstrate the practicality of their schemes. It allows insights regarding competitiveness and gives valuable hints in regard to optimization and parameter setting. A prior mock-up implementation was for example used by Bindel et al. (see [@10.1007/978-3-319-59879-6_22]) for a first evaluation of the hybrid certificate approach in conjunction with the qTESLA scheme. Our software allows a more thorough evaluation of such schemes.
+
 # Hybrid certificates
 
 Hybrid certificates [@10.1007/978-3-319-59879-6_22] are the basis for the use of hybrid cryptography. They are signed in parallel with two different signature schemes and additionally bind two independent public keys--one classical and one post-quantum key--to one identity. 
