### 2019-07-29 #1561 @j-braun
Hi @danielskatz  , just added some text regarding research applications to the paper as @kyleniemeyer  requested.
