### 2019-07-29 #1561 @danielskatz
👋  @msalle @DrDaveD @jteheran @zsshah & @djw8605 - would one of you be willing to review this submission for JOSS?  See the [paper above](https://github.com/openjournals/joss-papers/blob/joss.01561/joss.01561/10.21105.joss.01561.pdf) and [info on JOSS reviewing](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)
