### 2019-07-29 #1561 @danielskatz
Thanks @jteheran - it's an open interactive process, so there's not exactly a deadline,  but we like to have reviewers provide initial feedback in a couple of weeks, if possible
