### 2019-07-29 #1606 @danielskatz
👋 @jbasney, @jteheran - Here's where we do the review - please read the comments in this issue for instructions, but basically, your job is to either check off your items in the first comment or explain why you can't - either briefly here, or by opening issues in the source repo and mentioning this issue in them to create a link from here to there.

Thanks for being willing to do this, and if you have any questions, please ask me
