### 2019-08-05 CROSSINGTUD/bc-hybrid-certificates#1 @jbasney
v1.0.0 release
openjournals/joss-reviews#1606 references a v1.0.0 release for this repo. Please create a release tag so the v1.0.0 release appears at https://github.com/CROSSINGTUD/bc-hybrid-certificates/releases.
