### 2019-08-05 CROSSINGTUD/bc-hybrid-certificates#2 @jbasney
Add DOIs to JOSS PDF
In https://github.com/openjournals/joss-papers/blob/joss.01606/joss.01606/10.21105.joss.01606.pdf for openjournals/joss-reviews#1606 the first two references do not have URLs.

For the first one, I see the URL in the bibtex note in https://github.com/CROSSINGTUD/bc-hybrid-certificates/blob/master/paper/paper.bib so I think some bibtex magic is needed to make it appear in the JOSS PDF.

For the second one, I see the DOI in the bibtex entry name ([10.1007/978-3-319-59879-6_22](http://doi.org/10.1007/978-3-319-59879-6_22)) so I think that bibtex entry just needs an added `doi={10.1007/978-3-319-59879-6_22}` line so the DOI appears in the PDF.

Also, please generate a DOI for the v1.0.0 release (#1) (see: https://guides.github.com/activities/citable-code/) that you can include in the JOSS PDF.

The above will satisfy the JOSS requirement to have DOIs for the papers/software (https://joss.readthedocs.io/en/latest/review_checklist.html#software-paper).
