### 2019-08-09 CROSSINGTUD/bc-hybrid-certificates#2 @danielskatz
Actually, JOSS does not require DOIs for references - it requires that if a reference has a DOI, it should be part of the reference.
