### 2019-08-09 CROSSINGTUD/bc-hybrid-certificates#1 @j-braun
this has been fixed, I just added the tag.
