### 2019-08-09 13e86fb
fixed bib entries
