+++ paper/paper.bib
@@ -26,7 +26,8 @@ publisher="Springer International Publishing",
 address="Cham",
 pages="384--405",
 abstract="To ensure uninterrupted cryptographic security, it is important to begin planning the transition to post-quantum cryptography. In addition to creating post-quantum primitives, we must also plan how to adapt the cryptographic infrastructure for the transition, especially in scenarios such as public key infrastructures (PKIs) with many participants. The use of hybrids---multiple algorithms in parallel---will likely play a role during the transition for two reasons: ``hedging our bets'' when the security of newer primitives is not yet certain but the security of older primitives is already in question; and to achieve security and functionality both in post-quantum-aware and in a backwards-compatible way with not-yet-upgraded software.",
-isbn="978-3-319-59879-6"
+isbn="978-3-319-59879-6",
+doi="10.1007/978-3-319-59879-6_22"
 }
 
 @misc{OpenQuantumSafe,
@@ -84,5 +85,5 @@ month={Nov},}
   title        = {{qTESLA}},
   institution  = {{N}ational {I}nstitute of {S}tandards and {T}echnology},
   year         = 2019,
-  note         = {available at \url{https://csrc.nist.gov/CSRC/media/Projects/Post-Quantum-Cryptography/documents/round-2/submissions/qTESLA-Round2.zip}},
+  howpublished = {\url{https://csrc.nist.gov/CSRC/media/Projects/Post-Quantum-Cryptography/documents/round-2/submissions/qTESLA-Round2.zip}},
 }
