### 2019-08-09 CROSSINGTUD/bc-hybrid-certificates#2 @j-braun
I have fixed the bibentries so doi and URL ar shown now in the compiled Paper:
https://github.com/openjournals/joss-papers/blob/joss.01606/joss.01606/10.21105.joss.01606.pdf
Regarding the DOI for the v1.0.0 release I need to check with our Organizaition administrators whether its fine to grant organization access to Zenodo. I will come back to this issue once I have a reply.
