### 2019-08-09 f46241b
added doi for release v1.0.0
