+++ paper/paper.bib
@@ -49,7 +49,8 @@ quantum-resistant cryptography}},
   title = {{Hybrid Certificates - Java, Bouncy Castle integration}},
 	year="2019",
   howpublished = {\url{https://github.com/CROSSINGTUD/bc-hybrid-certificates}},
-  note = {Online; accessed 2019-07}
+  note = {Online; accessed 2019-07},
+	doi={10.5281/zenodo.3364471}
 }
 
 @misc{bouncyCastle,
