### 2019-08-09 CROSSINGTUD/bc-hybrid-certificates#2 @j-braun
I additionally requested a DOI for release v1.0.0 as suggested and added it to the paper.
