### 2019-08-09 #1606 @j-braun
@danielskatz  - we took care of the issues and fixed them
