### 2019-08-09 #1606 @danielskatz
@j-braun - please merge https://github.com/CROSSINGTUD/bc-hybrid-certificates/pull/4 or let me know any changes you don't agree with.
