### 2019-08-10 #1606 @j-braun
@danielskatz I agreed to your changes, and only fixed one last typo.
Version is now v1.0.1
DOI: 10.5281/zenodo.3365225

@danielskatz, @jbasney, @jteheran -  thank you very much for your quick review and support
