### 2019-08-12 #1606 @j-braun
@danielskatz - Updating the title in Zenodo was not possible. Instead Icreated a new release matching the paper title. This triggered a new archive, but also created a new DOI.
DOI: 10.5281/zenodo.3365786
hope this is ok.
